Intake DB structure ER diagram v0.1
This is the ER diagram of the database that I will be running on MySQL for my PC intake and processing of my computers that I have taken in. This is purely for PC type machines and does not include embedded systems. This is the first version and I anticipate expanding upon it as my lab evolves to more advanced technicalities.